move compression code to compress.c
