+++ src/backend/global.h
@@ -481,12 +481,14 @@ void rtlsym_init();
 void rtlsym_reset();
 void rtlsym_term();
 
+// compress.c
+char *id_compress(char *id, int idlen, size_t *plen);
+
 // Dwarf
 void dwarf_CFA_set_loc(size_t location);
 void dwarf_CFA_set_reg_offset(int reg, int offset);
 void dwarf_CFA_offset(int reg, int offset);
 void dwarf_CFA_args_size(size_t sz);
-
 
 #if TARGET_LINUX || TARGET_OSX || TARGET_FREEBSD || TARGET_OPENBSD || TARGET_SOLARIS
 elem * exp_isconst();
